add hack again

+++ crates/serverless/src/deployments/filesystem.rs
@@ -18,9 +18,13 @@ pub fn create_deployments_folder() -> Result<()> {
 pub fn rm_deployment(deployment_id: &str) -> Result<()> {
     #[cfg(not(test))]
     {
-        let path = Path::new(DEPLOYMENTS_DIR).join(deployment_id);
-        if path.exists() {
-            fs::remove_dir_all(path)?;
+        // TODO: use some compiler flag instead of env var
+        let skip = std::env::var("LAGOSS_TEST_SKIP_RM_DEPLOYMENT").is_ok();
+        if !skip {
+            let path = Path::new(DEPLOYMENTS_DIR).join(deployment_id);
+            if path.exists() {
+                fs::remove_dir_all(path)?;
+            }
         }
     }
